Fixes for Jade compat
