+++ src/main/java/twilightforest/compat/jade/DryingRackComponentProvider.java
@@ -3,15 +3,13 @@ package twilightforest.compat.jade;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.Identifier;
-import net.minecraft.world.level.block.entity.BlockEntity;
 import snownee.jade.api.*;
 import snownee.jade.api.config.IPluginConfig;
 import snownee.jade.api.ui.JadeUI;
 import twilightforest.TwilightForestMod;
-import twilightforest.block.entity.DryingRackBlockEntity;
 import twilightforest.compat.RecipeViewerConstants;
 
-public enum DryingRackProvider implements IBlockComponentProvider, IServerDataProvider<BlockAccessor> {
+public enum DryingRackComponentProvider implements IBlockComponentProvider {
 	INSTANCE;
 
 	@Override
@@ -21,16 +19,6 @@ public enum DryingRackProvider implements IBlockComponentProvider, IServerDataPr
 			int progress = data.getIntOr("progress", 0);
 			int total = data.getIntOr("total", 0);
 			tooltip.add(JadeUI.text(Component.translatable("jade.drying_rack.remaining", RecipeViewerConstants.getDryingTime(total - progress))).offset(10, 0));
-		}
-	}
-
-	@Override
-	public void appendServerData(CompoundTag data, BlockAccessor accessor) {
-		BlockEntity entity = accessor.getBlockEntity();
-		if (entity instanceof DryingRackBlockEntity rack && rack.isDrying()) {
-			CompoundTag tag = rack.saveWithoutMetadata(accessor.getLevel().registryAccess());
-			data.putInt("progress", tag.getIntOr("dry_time", 0));
-			data.putInt("total", tag.getIntOr("total_dry_time", 0));
 		}
 	}
