+++ src/main/java/twilightforest/compat/jade/QuestRamWoolProvider.java
@@ -17,7 +17,6 @@ import twilightforest.util.ColorUtil;
 import java.util.Map;
 
 public enum QuestRamWoolProvider implements IEntityComponentProvider {
-
 	INSTANCE;
 
 	@Override
